+++ tests/lib/Integration/Routing/HasManyTest.php
@@ -74,7 +74,8 @@ class HasManyTest extends TestCase
     public function test(string $method, string $uri, string $action, string $name): void
     {
         $server = $this->createServer('v1');
-        $this->createSchema($server, 'posts', '\d+');
+        $schema = $this->createSchema($server, 'posts', '\d+');
+        $this->createRelation($schema, 'tags');
 
         $this->defaultApiRoutesWithNamespace(function () {
             JsonApiRoute::server('v1')->prefix('v1')->namespace('Api\\V1')->resources(function ($server) {
@@ -99,13 +100,13 @@ class HasManyTest extends TestCase
      * @param string $method
      * @param string $uri
      * @param string $action
-     * @param string $name
      * @dataProvider genericProvider
      */
-    public function testName(string $method, string $uri, string $action, string $name): void
+    public function testName(string $method, string $uri, string $action): void
     {
         $server = $this->createServer('v1');
-        $this->createSchema($server, 'posts', '\d+');
+        $schema = $this->createSchema($server, 'posts', '\d+');
+        $this->createRelation($schema, 'tags');
 
         $this->defaultApiRoutesWithNamespace(function () use ($action) {
             JsonApiRoute::server('v1')
@@ -131,7 +132,8 @@ class HasManyTest extends TestCase
     public function testMiddleware(string $method, string $uri): void
     {
         $server = $this->createServer('v1');
-        $this->createSchema($server, 'posts', '\d+');
+        $schema = $this->createSchema($server, 'posts', '\d+');
+        $this->createRelation($schema, 'tags');
 
         $this->defaultApiRoutesWithNamespace(function () {
             JsonApiRoute::server('v1')
@@ -157,7 +159,8 @@ class HasManyTest extends TestCase
     public function testUri(string $method, string $uri): void
     {
         $server = $this->createServer('v1');
-        $this->createSchema($server, 'posts', '\d+');
+        $schema = $this->createSchema($server, 'posts', '\d+');
+        $this->createRelation($schema, 'blog-tags', 'tags');
 
         $this->defaultApiRoutesWithNamespace(function () {
             JsonApiRoute::server('v1')
@@ -165,7 +168,7 @@ class HasManyTest extends TestCase
                 ->namespace('Api\\V1')
                 ->resources(function ($server) {
                     $server->resource('posts')->relationships(function ($relations) {
-                        $relations->hasMany('blog-tags')->uri('tags');
+                        $relations->hasMany('blog-tags');
                     });
                 });
         });
@@ -234,7 +237,8 @@ class HasManyTest extends TestCase
     public function testOnly($only, array $matches): void
     {
         $server = $this->createServer('v1');
-        $this->createSchema($server, 'posts', '\d+');
+        $schema = $this->createSchema($server, 'posts', '\d+');
+        $this->createRelation($schema, 'tags');
 
         $this->defaultApiRoutesWithNamespace(function () use ($only) {
             JsonApiRoute::server('v1')
@@ -312,7 +316,8 @@ class HasManyTest extends TestCase
     public function testExcept($except, array $matches): void
     {
         $server = $this->createServer('v1');
-        $this->createSchema($server, 'posts', '\d+');
+        $schema = $this->createSchema($server, 'posts', '\d+');
+        $this->createRelation($schema, 'tags');
 
         $this->defaultApiRoutesWithNamespace(function () use ($except) {
             JsonApiRoute::server('v1')
@@ -331,7 +336,8 @@ class HasManyTest extends TestCase
     public function testReadOnly(): void
     {
         $server = $this->createServer('v1');
-        $this->createSchema($server, 'posts', '\d+');
+        $schema = $this->createSchema($server, 'posts', '\d+');
+        $this->createRelation($schema, 'tags');
 
         $this->defaultApiRoutesWithNamespace(function () {
             JsonApiRoute::server('v1')
@@ -403,7 +409,8 @@ class HasManyTest extends TestCase
     public function testOwnAction(string $method, string $uri, string $action, string $expected): void
     {
         $server = $this->createServer('v1');
-        $this->createSchema($server, 'posts', '\d+');
+        $schema = $this->createSchema($server, 'posts', '\d+');
+        $this->createRelation($schema, 'tags');
 
         $this->defaultApiRoutesWithNamespace(function () use ($action) {
             JsonApiRoute::server('v1')->prefix('v1')->namespace('Api\\V1')->resources(function ($server) use ($action) {
@@ -430,7 +437,8 @@ class HasManyTest extends TestCase
     public function testOwnActions(string $method, string $uri, string $action, string $expected): void
     {
         $server = $this->createServer('v1');
-        $this->createSchema($server, 'posts', '\d+');
+        $schema = $this->createSchema($server, 'posts', '\d+');
+        $this->createRelation($schema, 'tags');
 
         $this->defaultApiRoutesWithNamespace(function () {
             JsonApiRoute::server('v1')->prefix('v1')->namespace('Api\\V1')->resources(function ($server) {